fix: tests et ajout des suppression d'interaction

+++ packages/frontend/src/app/modules/usager-profil/components/profil-procuration-courrier/profil-procuration-courrier-component.ts
@@ -147,6 +147,11 @@ export class UsagersProfilProcurationCourrierComponent implements OnInit {
   }
 
   public deleteProcuration() {
+    if (!this.usager.options.procuration.actif) {
+      this.hideForm();
+      return;
+    }
+
     this.usagerProfilService.deleteProcuration(this.usager.ref).subscribe(
       (usager: UsagerLight) => {
         this.hideForm();
